requirements.txt
asgiref==3.2.3
Django==3.0
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-staticfiles==1.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
